TST: Move all the fetcher tests into one place.
